inventory/requirements.txt
aniso8601==1.0.0
Flask==0.10.1
Flask-MySQLdb==0.2.0
Flask-RESTful==0.3.4
gunicorn==18.0
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
requests==0.11.1
MySQL-python==1.2.5
mysqlclient==1.3.6
pytz==2015.4
six==1.9.0
Werkzeug==0.9.4
wheel==0.24.0
